Rename app folder to meetup_facebook_bot
