+++ .travis.yml
@@ -6,6 +6,6 @@ install:
   - pip3 install -r requirements.txt
   - pip3 install codecov
 script: 
-  - coverage run -m pytest tests --cov app
+  - coverage run -m pytest tests --cov meetup_facebook_bot
 after_success:
   - codecov
