+++ set_start_button.py
@@ -1,7 +1,9 @@
 import os
 
-from app import messenger
+from meetup_facebook_bot import messenger
 
 
-messenger.messenger_profile.set_get_started_button(os.environ['ACCESS_TOKEN'], 
-                                                   'get started payload')
+messenger.messenger_profile.set_get_started_button(
+    os.environ['ACCESS_TOKEN'], 
+   'get started payload'
+)
